+++ src/bot/CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - `BotOptions.recognizePrivateMessages`
 
+### Fixed
+
+- `lexCommandInput` now throws custom `SyntaxError` instead of built-in
+  `SyntaxError`.
+
 ## 0.1.1
 
 ### Fixed